+++ daemon/graphdriver/overlay2/overlay.go
@@ -180,7 +180,7 @@ func Init(home string, options []string, uidMaps, gidMaps []idtools.IDMap) (grap
 		if opts.overrideKernelCheck {
 			logger.Warn("Using pre-4.0.0 kernel for overlay2, mount failures may require kernel update")
 		} else {
-			if err := supportsMultipleLowerDir(testdir); err != nil {
+			if err := overlayutils.SupportsMultipleLowerDir(testdir); err != nil {
 				logger.Debugf("Multiple lower dirs not supported: %v", err)
 				return nil, graphdriver.ErrNotSupported
 			}
